netloc: new scheme for scotch
We use two types of architecture graphs for the mapping with scotch. The
first one is the architecture graph with only nodes. Then, we build an
architecture graph representing a node, with the cores. We do that
because if the topology of the network is a torus and a tree within the
nodes, we cannot have a single architecture graph in scotch.