netloc: fails if not same number of cores per node
When a mapping is computed by scotch, we need the same number of cores
by nodes because we first build a tree with only nodes (and not cores),
and then we use another tree for each node. So, because we cannot put a
wieght for each node, the mapping will be unbalanced, and to avoid that,
we do not allow unbalanced nodes.