doc/Makefile: cleanup the building of the readme
Explicitly depend on $(DOX_HTML_DIR)/index.html so that
make can take it either from srcdir or builddir.
Create $(DOX_HTML_DIR) in case we're building the README
in a builddir while using the prebuilt doxygen in srcdir.

This is useful for #526 where they build outside of srcdir
and doxygen-doc/hwloc.tag is more recent than the README.
It triggers a rebuild of the README, which failed before
because things were missing in the builddir.

Closes #526.

By the way, rename DOX_HTML_README with _TMP since it's
really a temp file.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>