linux: add NVDIMM subtype to NUMA nodes when relevant
When a NUMA node comes from a DAXDevice, look under the "nd" sysfs class
to check the devtype of the nmem that build the corresponding region.

In practive, we already had the DAXDevice info attr which almost meant
NVDIMM except when asking the kernel to expose part of its DRAM as non-volatile.

Signed-off-by: Brice Goglin <Brice.Goglin@inria.fr>